% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anscombe-documentation.R
\docType{data}
\name{anscombe}
\alias{anscombe}
\title{Anscombe Datasets}
\format{A data frame with 44 rows and 3 variables:
\describe{
  \item{dataset}{a character vector identifying each of the four datasets}
  \item{x}{x}
  \item{y}{y}
}}
\usage{
anscombe
}
\description{
Anscombe Datasets
}
\examples{
data(anscombe)
summary(anscombe)
}
\references{
Anscombe, Francis J. 1973. "Graphs in Statistical Analysis." *American Statistician*. 27(1): 17–21.

Tufte, Edward R. 2001. The Visual Display of Quantitative Information (2nd ed.). Cheshire, CT: Graphics Press.
}
\keyword{datasets}
