% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/therms-documentation.R
\docType{data}
\name{therms}
\alias{therms}
\title{ANES Feeling Thermometers, 1978-2016}
\format{A data frame with 38,100 rows and 4 variables:
\describe{
  \item{year}{the year of the survey}
  \item{education}{the respondent's level of education}
  \item{ft_democratic_party}{the respondent's feelings toward the Democratic party}
  \item{ft_republican_party}{the respondent's feelings toward the Democratic party}
}}
\usage{
therms
}
\description{
A tidied dataset from
\href{https://electionstudies.org/data-center/anes-time-series-cumulative-data-file/}{ANES}. This
dataset include variables for year, education, and feeling thermometers for the Republican and Democratic parties.
}
\examples{
# load data
data(therms)

plot(therms$ft_democratic_party, therms$ft_republican_party)
}
\keyword{datasets}
